requirements.txt
streamlit
requests
folium
streamlit-folium
streamlit-autorefresh
earthengine-api>=0.1.326
geemap>=0.25
google-auth